Renamed the scripts
